Added an initial check if the player has any food. This also guarantees that the 'food' key exists for the rest of the code in this method.
Reduced the try block to cover only the line with the int() call.

Added a check that the player doesn't enter a negative number of food to eat.

Let player enter 0 for amount, just to cancel the food eating.

Made the check for "don't have that much food" check earlier, so the inventory/health/hunger change doesn't have to be later undone.

Simplified the "cap health to max health" code.

Moved the stats-changing code outside of the while loop, since the loop is only needed for input validation.